Move policy protobuf handling in the `server-policy` crate
In preparation for additional policy types, this change policy protobuf
handling into the `server-policy` crate, gated by a `proto` feature.

This change introduces proper error types that enumerate the ways that
protobuf handling can go wrong.
