+++ linkerd/server-policy/Cargo.toml
@@ -6,8 +6,13 @@ license = "Apache-2.0"
 edition = "2021"
 publish = false
 
+[features]
+proto = ["linkerd2-proxy-api/inbound"]
+
 [dependencies]
 ipnet = "2"
+linkerd2-proxy-api = { version = "0.5", features = ["inbound"], optional = true }
+thiserror = "1"
 
 [dev-dependencies]
 quickcheck = { version = "1", default-features = false }
