+++ linkerd/server-policy/src/lib.rs
@@ -1,10 +1,12 @@
 #![deny(rust_2018_idioms, clippy::disallowed_methods, clippy::disallowed_types)]
 #![forbid(unsafe_code)]
 
-mod network;
+pub mod authz;
+#[cfg(feature = "proto")]
+mod proto;
 
-pub use self::network::Network;
-use std::{borrow::Cow, collections::HashSet, hash::Hash, sync::Arc, time};
+pub use self::authz::{Authentication, Authorization};
+use std::{borrow::Cow, hash::Hash, sync::Arc, time};
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct ServerPolicy {
@@ -23,95 +25,9 @@ pub enum Protocol {
     Tls,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Authorization {
-    pub networks: Vec<Network>,
-    pub authentication: Authentication,
-    pub meta: Arc<Meta>,
-}
-
 #[derive(Debug, PartialEq, Eq, Hash)]
 pub struct Meta {
     pub group: Cow<'static, str>,
     pub kind: Cow<'static, str>,
     pub name: Cow<'static, str>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum Authentication {
-    Unauthenticated,
-    TlsUnauthenticated,
-    TlsAuthenticated {
-        identities: HashSet<String>,
-        suffixes: Vec<Suffix>,
-    },
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Suffix {
-    ends_with: String,
-}
-
-// === impl Suffix ===
-
-impl From<Vec<String>> for Suffix {
-    fn from(parts: Vec<String>) -> Self {
-        let ends_with = if parts.is_empty() {
-            "".to_string()
-        } else {
-            format!(".{}", parts.join("."))
-        };
-        Suffix { ends_with }
-    }
-}
-
-impl Suffix {
-    #[inline]
-    pub fn contains(&self, name: &str) -> bool {
-        name.ends_with(&self.ends_with)
-    }
-}
-
-#[cfg(test)]
-mod network_tests {
-    use super::Network;
-    use ipnet::{IpNet, Ipv4Net, Ipv6Net};
-    use quickcheck::{quickcheck, TestResult};
-    use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
-
-    quickcheck! {
-        fn contains_v4(addr: Ipv4Addr, exclude: Option<Ipv4Addr>) -> TestResult {
-            let net = Network {
-                net: Ipv4Net::default().into(),
-                except: exclude.into_iter().map(|a| IpNet::from(IpAddr::V4(a))).collect(),
-            };
-
-            if let Some(e) = exclude {
-                if net.contains(&e.into()) {
-                    return TestResult::failed();
-                }
-                if addr == e {
-                    return TestResult::passed();
-                }
-            }
-            TestResult::from_bool(net.contains(&addr.into()))
-        }
-
-        fn contains_v6(addr: Ipv6Addr, exclude: Option<Ipv6Addr>) -> TestResult {
-            let net = Network {
-                net: Ipv6Net::default().into(),
-                except: exclude.into_iter().map(|a| IpNet::from(IpAddr::V6(a))).collect(),
-            };
-
-            if let Some(e) = exclude {
-                if net.contains(&e.into()) {
-                    return TestResult::failed();
-                }
-                if addr == e {
-                    return TestResult::passed();
-                }
-            }
-            TestResult::from_bool(net.contains(&addr.into()))
-        }
-    }
 }
